moved dom manipulation functions to module
