+++ client/package.json
@@ -20,6 +20,7 @@
     "gulp-util": "^2.2.14",
     "gulp-webpack": "^1.5.0",
     "html-loader": "^0.4.0",
+    "image-webpack-loader": "^1.6.2",
     "imports-loader": "^0.6.5",
     "node-sass": "^3.4.2",
     "sass-loader": "^3.1.2",
