+++ client/www/assets/scss/_wizard.scss
@@ -1,0 +1,17 @@
+.enemy {
+  margin-bottom: 10%;
+}
+.avatar-container {
+  position: relative;
+}
+
+.avatar-overlay {
+  visibility: hidden;
+  position: absolute;
+  top: 250px;
+  background-color: $black;
+  color: ivory;
+  text-align: center;
+  width: 50%;
+  opacity: 0.75;
+}
