+++ client/www/views/components/wizard/wizard.html
@@ -1,30 +1,34 @@
-<div id="{{wizard.id}}">
+<div id="{{wizard.id}}" class="enemy">
   <div class="row">
-    <div class="col">
-      <h3>{{ wizard.user }}</h3>
-    </div>
-  </div>
-  <div class="row">
-    <div class="col" class="avatar-container">
-      <img style="max-width: 75%; width: auto; height:auto" src="{{ wizard.avatar }}" alt="Avatar for {{ wizard.user }}" />
-      <div id="{{wizard.id}}-avatar-overlay" class="avatar-overlay">
-        <span id="{{wizard.id}}-wizard-message" class="wizard-message"></span>
+    <div class="col col-33 col-offset-33">
+      <div class="row">
+        <div class="col">
+          <span id="{{wizard.id}}-health" class="health-bar" style="width:{{health}}%"></span>
+        </div>
       </div>
-    </div>
-  </div>
-  <div class="row">
-    <div class="col">
-      <span id="{{wizard.id}}-health" class="health-bar" style="width:{{health}}%"></span>
-    </div>
-  </div>
-  <div class="row">
-    <div class="col">
-      <span id="{{wizard.id}}-mana" class="mana-bar" style="width:{{mana}}%"></span>
-    </div>
-  </div>
-  <div class="row">
-    <div class="col">
-      <span id="{{wizard.id}}-spell-message" class="spell-message"></span>
+      <div class="row">
+        <div class="col">
+          <span id="{{wizard.id}}-mana" class="mana-bar" style="width:{{mana}}%"></span>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col">
+          <span id="{{wizard.id}}-spell-message" class="spell-message"></span>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col">
+          <h3>{{ wizard.user }}</h3>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col" class="avatar-container">
+          <img style="max-width: 75%; width: auto; height:auto" src="{{ wizard.avatar }}" alt="Avatar for {{ wizard.user }}" />
+          <div id="{{wizard.id}}-avatar-overlay" class="avatar-overlay">
+            <span id="{{wizard.id}}-wizard-message" class="wizard-message"></span>
+          </div>
+        </div>
+      </div>
     </div>
   </div>
 </div>
